Removed CLI flag registration from liaison and moved CLI registration.
We're actually going to take a different approach that doesn't require
access to the Docker CLI flags: we're going to pull the Docker host
information off of the client established by the CLI. This will be far
simpler and more consistent. It should allow us to simplify the Mutagen
Docker transport as well since we can now just embed this as a
DOCKER_HOST environment variable (as opposed to storing flags).

This commit also relocates registration of the Docker CLI instance to
reside next to the rest of the liaison registration code.

Signed-off-by: Jacob Howard <jacob@mutagen.io>