Added check for --no-deps flag in run.
If the --no-deps flag is specified to run and the target service has a
dependency on the Mutagen Compose sidecar service, then we need to
display an error because (even though the Mutagen Compose sidecar
service will be started given our project manipulation) the underlying
Compose implementation won't wait for the sidecar to start.

We don't need to handle the --no-deps flag on up because it behaves a
little differently, and our project manipulation will ensure that the
Mutagen Compose sidecar service still starts with all dependent services
waiting on it appropriately.

Signed-off-by: Jacob Howard <jacob@mutagen.io>