Integrated Mutagen Compose status updates with native Compose output.
Mutagen Compose will now write (most) of its status updates via the
Compose progress package. A few things remain to fix, notably the output
that currently comes from Mutagen's prompters in the *WithSpecification
and *WithSelection functions. We'll need to modify those to accept a
custom prompter (though fortunately we've already implemented the
necessary interface into the new statusUpdater type).

We also need to figure out how to get more reconcileSessions output to
show up, because by default up won't display status updates if
attaching, and that means most of that Mutagen status information is
lost to the log output. We'll likely need to bring Mutagen up first (and
start it detached) to get that output.

Signed-off-by: Jacob Howard <jacob@mutagen.io>