Modified up, create, start, and run to operate on Mutagen first.
By default (and in most usage) start progress updates (such as those
emitted by Liaison.reconcileSessions) are hidden because they would
conflict with log output. This is a problem because that's where
Mutagen's most important updates originate. To solve this issue, this
commit modifies up (and analagous commands) to operate on Mutagen before
other services. This provides an opportunity to have these progress
updates displayed and to enforce consistency with how the Mutagen
service is created and started.

There is a hypothetical loss of some efficiency here vs. the
depends_on-based startup of Mutagen, but the added value of consistency
and visibility is worth the tradeoff, especially since 99% of use cases
would not benefit from the depends_on-based startup and the 1% that
would can just docker compose commands.

Signed-off-by: Jacob Howard <jacob@mutagen.io>