Added build script based on Mutagen's build script.
This commit adds a modified version of the Mutagen build script,
tailored for Mutagen Compose. Ideally we'll eventually extract and
export most of the infrastructure from the Mutagen build script into a
separate package that we can import here.

This script doesn't currently support all Mutagen platforms, as many
require patches for Docker dependencies, but we do support macOS,
Windows, and Linux, and we can expand this going forward.

Signed-off-by: Jacob Howard <jacob@mutagen.io>