Added license printing infrastructure.
This commit adds the "legal" command for displaying license information
(as well as any other legal information that might be necessary). We
could probably call this "licenses" instead, but "legal" maintains
consistency with Mutagen and is more general.

This commit also adds initial license text, but we still need to add
license text for all dependencies.

Signed-off-by: Jacob Howard <jacob@mutagen.io>