Removed service list modifications in RunOneOffContainer.
Given that the run command will always invoke Create and Start before
invoking RunOneOffContainer (even if there are no explicit dependencies,
most likely to create volumes/networks), we don't need to bring up the
Mutagen service manually in RunOneOffContainer. Obviously this is code
that we'll need to monitor for changes when rebasing in the future.

Signed-off-by: Jacob Howard <jacob@mutagen.io>