start: only reconcile sessions if the project has been processed
When invoking start indirectly via up, we'll have session definitions
that are loaded when processing the project, but if invoking start
directly, then the session definitions will be empty, and thus we don't
want reconcileSessions to remove all existing sessions. However, in both
cases we still want the sessions running, so we'll just start whatever's
there if we don't have session definitions. This aligns significantly
better with Compose's natural behavior.

Signed-off-by: Jacob Howard <jacob@mutagen.io>