run: ensure sidecar service is always started on run
In github.com/docker/compose@dbe7de50a (first shipped in Compose
v2.3.1), a restriction was added where the backend's Start method would
only be invoked if the run target had a non-zero number of dependencies.
This meant that the Mutagen sidecar service often wasn't started when
run was invoked. To fix this, we'll track Start's invocation and perform
the start operation in RunOneOffContainer if necessary. A more holistic
solution is probably warranted, but it'll have to wait until the
internal Compose APIs stabilize a bit more.

Signed-off-by: Jacob Howard <jacob@mutagen.io>