cli: add support for command line completion
This commit adds a generate command to create shell completion scripts
and makes the CLI entry point aware of Cobra's completion commands.

Unfortunately, our argument emulation doesn't allow completion to work
for top-level flag values (though it will work for those flags' names
(which don't require invoking Go-based completion), as well as for
subcommand names, flags, values, and arguments). The reason is that we
have to parse and separate top-level flags before dispatching the
completion request to the plugin infrastructure, which we can't do if
the flag isn't fully specified. We could try to come up with a looser,
more name-aware heuristic for separating these flags, but it would be
incredibly difficult, and it doesn't seem as if the standalone Docker
Compose v2 Convert function makes this effort; it also rejects
top-level flags without values when performing completion.

Adjustments will still need to be made to the Homebrew formula to take
advantage of this completion.

Signed-off-by: Jacob Howard <jacob@mutagen.io>