all: remove MIT-licensed sidecar targeting
We're no longer going to ship the MIT-licensed sidecar image starting
with Mutagen v0.17, so we can remove the ability to target it. We'll
retain support for the corresponding configuration field, but it no
longer has any effect.

For the time being, we're not going to switch the build of Mutagen
Compose itself to include SSPL-licensed enhancements, but we may change
this in the future if required. At the moment, Mutagen Compose only
generates session configurations, so it doesn't need the corresponding
SSPL-licensed code to take advantage of those configurations, but if we
start doing configuration validation, or something else that requires
the SSPL-licensed support code, then we may have to enable that code.

Signed-off-by: Jacob Howard <jacob@mutagen.io>