all: sync build script cgo behavior with Mutagen
Since we've now enabled cgo for Mutagen Compose builds, we also need to
add in the minimum macOS version targeting required for portability.
Moreover, we need to disable cgo if trying to cross-compile from a
non-Darwin platform and warn if this is an issue.

Note that this change hides the FSEventStreamScheduleWithRunLoop
deprecation warning since that function isn't deprecated on the macOS
versions that we're targeting.

Signed-off-by: Jacob Howard <jacob@mutagen.io>