sidecar: disable networking if no network dependencies are present
The newer version of Compose used with Mutagen Compose v0.18.0 doesn't
like an empty network map and tries to connect the sidecar container to
a network with an empty name if there are no networks present in the
service configuration.

To fix this, we'll disable networking for the sidecar if not needed.

Signed-off-by: Jacob Howard <jacob@mutagen.io>